pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anti-grav"
version = "0.1.0"
description = "News Scraper and Analyzer"
authors = [
  { name="User", email="user@example.com" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "streamlit",
    "playwright",
    "beautifulsoup4",
    "lxml",
    "newspaper3k",
    "textblob",
    "sqlalchemy",
    "pandas",
    "plotly",
    "lxml_html_clean",
    "fpdf"
]

[project.urls]
"Homepage" = "https://github.com/example/anti-grav"
